requirements.txt
url_normalize
tgcrypto
httpx
APScheduler
git+https://github.com/KurimuzonAkuma/pyrogram.git@v2.1.17
python-dotenv
flask
psycopg2
geopy